fix GCC-10 / -fno-common build. Fix #7
Bug was originally reported at Gentoo Bugzilla as Bug 706692 by Toralf Förster
on 2020-01-27, and Matt Whitlock on 2020-05-16 provided the fix used in this
commit, and as such is credited as commit author. On 2020-07-04 Jonas Stein
mentioned this repository on that Gentoo Bug Report and reported the issue on
Github.

See https://bugs.gentoo.org/706692

Thanks Jonas Stein @jonasstein for reporting!