avoid GCC warning: comparison between pointer and zero character constant
The puzzling comparison `if (file->filename == '\0') ...` appeared in several
parts of the code, mostly in main.c@process_files(). It issued the above warning
in modern GCC (tested with 7.5.0). It is not clear if the original author's
intention was to test if the filename _content_ was '\0' (ie, a blank filename)
or the filename pointer _itself_ was '\0', which seems a very weird way to test
for a NULL pointer.

Given that on a particular line it sets `file->filename = '\0'`, it is assumed
the intention was to point filename to NULL, specially since filename's content
is a const.

So, all comparisions GCC warned about were changed to `if (file->filename == NULL) ...`,
and that assignment changed to `file->filename = NULL`. It should be noted that there
might be other similar (wrong) comparisons that GCC didn't find/complained and, most
importantly, similar assignments other than the one fixed.